+++ lib/index.ts
@@ -1,134 +1,156 @@
+/* eslint-disable complexity */
+/* eslint-disable sonarjs/cognitive-complexity */
 /* eslint-disable no-use-before-define */
 /* eslint-disable no-unused-vars */
-type VnodeOrUnknown = unknown[] | ComponentVnode | Vnode | TextVnode | unknown | VnodeOrUnknown[];
 
-type DomElement = (Element | SVGElement | HTMLElement) & Record<string, unknown>;
+type VnodeOrUnknown = VnodeComponent | Vnode | TextVnode | any;
 
 type DomAttribute = { nodeName: string; nodeValue: string };
 
-export type Props = Record<string, unknown> & {
+type DomElement = (HTMLElement | SVGElement) & Record<string, any>;
+
+type Props = {
   key?: string | number;
   data?: string;
   oncreate?: { (vnode: Vnode): never };
   onupdate?: { (vnode: Vnode, oldVnode: Vnode | TextVnode): never };
   onremove?: { (oldVnode: Vnode): never };
   onbeforeupdate?: { (vnode: Vnode, oldVnode: Vnode | TextVnode): undefined | boolean };
-};
+} & Record<string, any>;
 
-type Component = (props?: Record<string, unknown> | null, children?: VnodeOrUnknown) => VnodeOrUnknown | VnodeOrUnknown[];
+type Component = (props?: Record<string, any> | null, children?: VnodeOrUnknown) => VnodeOrUnknown | VnodeOrUnknown[];
 
-export type ValyrianComponent =
+type ValyrianComponent =
   | Component
-  | (Record<string, unknown> & {
+  | (Record<string, any> & {
       view: Component;
     });
 
-export interface Valyrian {
-  (tagOrComponent: string | ValyrianComponent, props?: Props | null, children?: VnodeOrUnknown): Vnode | ComponentVnode;
-  isMounted: boolean;
-  isNode: boolean;
-  reservedWords: string[];
-  current: Current;
-  usePlugin: (plugin: Plugin, options: Record<string, unknown>) => void;
-  trust: (htmlString: string) => Vnode[];
-  onCleanup: (callback: typeof Function) => void;
-  updateProperty: (name: string, newNode: Vnode, oldNode?: Vnode) => void;
-  update: (props?: Props, ...children: VnodeOrUnknown[]) => string | boolean | void;
-  mount: (container: string, component: ValyrianComponent, props?: Props | null, ...children: VnodeOrUnknown[]) => string | boolean | void;
-  unMount: () => string | boolean | void;
-  directive: (directive: string, handler: Directive) => void;
-  newInstance: () => Valyrian;
+type Current = { parentVnode?: Vnode; oldParentVnode?: Vnode; component?: VnodeComponent };
+
+interface Plugin {
+  (v: Valyrian, options?: Record<string, any>): never;
 }
 
-export interface Directive {
-  (value: any, vnode: Vnode, oldVnode?: Vnode | TextVnode): unknown;
+interface Directive {
+  (value: any, vnode: Vnode, oldVnode?: Vnode | TextVnode): void | boolean;
 }
 
-export interface Plugin {
-  (v: Valyrian, options: Record<string, unknown>): unknown;
-}
-
-export type Current = { parentVnode: Vnode | undefined; oldParentVnode: Vnode | undefined; component: ComponentVnode | undefined };
-
-export class Vnode {
+interface Vnode {
   name: string;
   props: Props;
   children: VnodeOrUnknown[];
   dom?: DomElement;
-  onCleanup?: unknown;
+  onCleanup?: FunctionConstructor[];
+  isSVG?: boolean;
+  processed?: boolean;
+}
+
+class Vnode implements Vnode {
+  name: string;
+  props: Props;
+  children: VnodeOrUnknown[];
+  dom?: DomElement;
+  onCleanup?: FunctionConstructor[];
   isSVG?: boolean;
   processed?: boolean;
 
-  constructor(name: string, props: Props | null = null, children: VnodeOrUnknown[]) {
-    this.props = props || {};
+  constructor(name: string, props: Props, children: VnodeOrUnknown) {
+    this.props = props;
     this.children = children;
     this.name = name;
   }
 }
 
-export class TextVnode {
-  dom?: DomElement;
+interface TextVnode {
+  dom?: Text;
+  nodeValue: string;
+}
 
-  constructor(dom?: DomElement) {
-    this.dom = dom;
+class TextVnode implements TextVnode {
+  dom?: Text;
+  nodeValue: string;
+
+  constructor(nodeValue: string) {
+    this.nodeValue = nodeValue;
   }
 }
 
-export class ComponentVnode {
+interface VnodeComponent {
   component: ValyrianComponent;
-  props: Props | null;
-  children: VnodeOrUnknown;
+  props: Props;
+  children: VnodeOrUnknown[];
+}
 
-  constructor(component: ValyrianComponent, props: Props | null = null, children: VnodeOrUnknown) {
+class VnodeComponent implements VnodeComponent {
+  component: ValyrianComponent;
+  props: Props;
+  children: VnodeOrUnknown[];
+
+  constructor(component: ValyrianComponent, props: Props, children: VnodeOrUnknown[]) {
     this.props = props;
     this.children = children;
     this.component = component;
   }
 }
 
-const UND: undefined = undefined;
-const NULL = null;
-const isArray = Array.isArray;
-const functionstr = "function";
-const once = "v-once";
-const key = "key";
-const svg = "svg";
-const str = "string";
+interface Valyrian {
+  (tagOrComponent: string | ValyrianComponent, props?: Props | null, children?: VnodeOrUnknown): Vnode | VnodeComponent;
+  isMounted: boolean;
+  isNode: boolean;
+  reservedWords: string[];
+  current: Current;
+  trust: (htmlString: string) => Vnode[];
+  usePlugin: (plugin: Plugin, options: Record<string, any>) => void;
+  onCleanup: (callback: typeof Function) => void;
+  updateProperty: (name: string, newVnode: Vnode & { dom: DomElement }, oldNode: Vnode & { dom: DomElement }) => void;
+  update: (props?: Props | null, ...children: VnodeOrUnknown) => string | void;
+  mount: (container: string | DomElement, component: ValyrianComponent, props?: Props | null, ...children: VnodeOrUnknown[]) => string | void;
+  unMount: () => string | boolean | void;
+  directive: (directive: string, handler: Directive) => void;
+  newInstance: () => Valyrian;
+  [x: string]: any;
+}
 
-const createElement = (tag: string, isSVG: boolean = false): DomElement =>
-  isSVG ? (document.createElementNS("http://www.w3.org/2000/svg", tag) as DomElement) : (document.createElement(tag) as DomElement);
+let isNode = typeof window === "undefined" || typeof global !== "undefined";
 
-// Hydrates the current dom before mount
-const domToVnode = (dom: DomElement) => {
-  if (dom.nodeType === 1) {
-    let props = {};
-    [].forEach.call(dom.attributes, (prop) => ((props as Props)[(prop as DomAttribute).nodeName] = (prop as DomAttribute).nodeValue));
+// Create Node element
+function createElement(tagName: string, isSVG: boolean = false): DomElement {
+  return isSVG ? document.createElementNS("http://www.w3.org/2000/svg", tagName) : document.createElement(tagName);
+}
 
-    let vnode: Vnode = new Vnode(dom.nodeName as string, props, []);
-    vnode.dom = dom;
+// Transforms a DOM node to a VNode
+function domToVnode(dom: DomElement): Vnode & { dom: DomElement } {
+  const el = dom;
+  const nodeName = el.nodeName.toLowerCase();
+  const props: Props = {};
+  const children: VnodeOrUnknown[] = [];
+  let i: number;
+  let childNode;
+  let attr: Attr;
 
-    for (let i = 0, l = dom.childNodes.length; i < l; i++) {
-      let childVnode = domToVnode(dom.childNodes[i] as DomElement);
-      childVnode && vnode.children.push(childVnode);
+  // attributes
+  for (i = 0; i < el.attributes.length; i++) {
+    attr = el.attributes[i];
+    props[attr.nodeName] = attr.nodeValue;
+  }
+
+  // children
+  for (i = 0; i < el.childNodes.length; i++) {
+    childNode = el.childNodes[i];
+    if (childNode.nodeType === 1) {
+      children.push(domToVnode(childNode as DomElement));
+    } else if (childNode.nodeType === 3) {
+      let textVnode = new TextVnode(childNode.nodeValue || "");
+      textVnode.dom = childNode as Text;
+      children.push(textVnode);
     }
-    return vnode;
-  }
-  if (dom.nodeType === 3) {
-    return new TextVnode(dom);
-  }
-};
-
-const emptyNode = new Vnode("empty", NULL, []);
-
-const callRemove = (vnode: Vnode) => {
-  for (let i = 0, l = vnode.children.length; i < l; i++) {
-    vnode.children[i] instanceof Vnode && callRemove(vnode.children[i] as Vnode);
   }
 
-  vnode.props.onremove && vnode.props.onremove(vnode);
-};
-
-const isNode = typeof window === "undefined" || typeof global !== "undefined";
+  let newNode = new Vnode(nodeName, props, children);
+  newNode.dom = dom;
+  return newNode as Vnode & { dom: DomElement };
+}
 
 const trust = (htmlString: string) => {
   let div = createElement("div");
@@ -137,43 +159,64 @@ const trust = (htmlString: string) => {
   return [].map.call(div.childNodes, (item) => domToVnode(item)) as Vnode[];
 };
 
-//eslint-disable-next-line max-lines-per-function
+// eslint-disable-next-line max-lines-per-function
 function valyrian(): Valyrian {
-  function v(tagOrComponent: string | ValyrianComponent, props: Props | null = null, ...children: VnodeOrUnknown[]) {
-    if (typeof tagOrComponent === str) {
-      return new Vnode(tagOrComponent as string, props, children);
+  const v: Valyrian = (tagOrComponent, props, ...children) => {
+    if (typeof tagOrComponent === "string") {
+      return new Vnode(tagOrComponent, props || {}, children);
+    } else {
+      return new VnodeComponent(tagOrComponent, props || {}, children);
     }
-
-    return new ComponentVnode(tagOrComponent as ValyrianComponent, props, children);
-  }
+  };
 
   v.isMounted = false;
-
   v.isNode = isNode;
-
-  let mainContainer: DomElement | null = NULL;
-  let mainNode: Vnode;
-  let oldMainNode: Vnode;
-  let mountedComponent: ValyrianComponent;
-
-  const reservedWords: string[] = [key, "data", once, "oncreate", "onupdate", "onremove", "onbeforeupdate"];
-  v.reservedWords = reservedWords;
+  v.reservedWords = ["key", "data", "v-once", "oncreate", "onupdate", "onremove", "onbeforeupdate"];
+  v.trust = trust;
 
   const current: Current = {
-    parentVnode: UND,
-    oldParentVnode: UND,
-    component: UND
+    parentVnode: undefined,
+    oldParentVnode: undefined,
+    component: undefined
   };
   v.current = current;
 
-  // Plugin system
   const plugins = new Map();
-  v.usePlugin = (plugin: Plugin, options: Record<string, unknown> = {}) => !plugins.has(plugin) && plugins.set(plugin, true) && plugin(v as Valyrian, options);
 
-  let attachedListeners: { [x: string]: boolean } = {};
+  v.usePlugin = (plugin: Plugin, options: Record<string, any> = {}) => !plugins.has(plugin) && plugins.set(plugin, true) && plugin(v as Valyrian, options);
+
+  let vnodesToCleanup: Vnode[] = [];
+
+  v.onCleanup = (callback: FunctionConstructor) => {
+    let parentVnode = v.current.parentVnode as Vnode;
+    if (!parentVnode.onCleanup) {
+      parentVnode.onCleanup = [] as FunctionConstructor[];
+    }
+
+    parentVnode.onCleanup.push(callback);
+
+    if (vnodesToCleanup.indexOf(parentVnode) === -1) {
+      vnodesToCleanup.push(parentVnode);
+    }
+  };
+
+  let cleanupVnodes = () => {
+    for (let l = vnodesToCleanup.length; l--; ) {
+      for (let callback of vnodesToCleanup[l].onCleanup as FunctionConstructor[]) {
+        callback();
+      }
+    }
+    vnodesToCleanup = [];
+  };
+
+  let mainContainer: DomElement | null = null;
+  let emptyComponent: ValyrianComponent = () => "";
+  let mountedComponent: ValyrianComponent = emptyComponent;
+
+  const attachedListeners: string[] = [];
   function eventListener(e: Event) {
     let dom = e.target as DomElement;
-    let name = `__on${e.type}`;
+    let name = `v-on${e.type}`;
     while (dom) {
       if (dom[name]) {
         (dom[name] as (a: Event, dom: DomElement) => void)(e, dom);
@@ -186,189 +229,145 @@ function valyrian(): Valyrian {
     }
   }
 
-  v.trust = trust;
+  function updateProperty(prop: string, newVnode: Vnode & { dom: DomElement }, oldVnode?: Vnode): void | boolean {
+    if (v.reservedWords.indexOf(prop) !== -1) {
+      if (directives[prop]) {
+        return directives[prop](newVnode.props[prop], newVnode, oldVnode);
+      }
+    } else if (typeof newVnode.props[prop] === "function" && newVnode.dom[`v-${prop}`] === undefined) {
+      if (attachedListeners.indexOf(prop) === -1) {
+        (mainContainer as DomElement).addEventListener(prop.slice(2), eventListener);
+        attachedListeners.push(prop);
+      }
 
-  let vnodesToCleanup: Vnode[] = [];
-
-  v.onCleanup = (callback: typeof Function) => {
-    let parentVnode = v.current.parentVnode as Vnode;
-    if (!parentVnode.onCleanup) {
-      parentVnode.onCleanup = [];
-    }
-
-    (parentVnode.onCleanup as FunctionConstructor[]).push(callback);
-
-    if (vnodesToCleanup.indexOf(parentVnode) === -1) {
-      vnodesToCleanup.push(parentVnode);
-    }
-  };
-
-  let cleanupVnodes = () => {
-    for (let l = vnodesToCleanup.length; l--; ) {
-      for (let callback of (vnodesToCleanup[l] as Vnode & { onCleanup: any }).onCleanup) {
-        callback();
+      newVnode.dom[`v-${prop}`] = newVnode.props[prop];
+    } else if (prop in newVnode.dom && !newVnode.isSVG) {
+      // eslint-disable-next-line eqeqeq
+      if (newVnode.dom[prop] != newVnode.props[prop]) {
+        newVnode.dom[prop] = newVnode.props[prop];
+      }
+    } else if (oldVnode === undefined || newVnode.props[prop] !== oldVnode.props[prop]) {
+      if (newVnode.props[prop] === false) {
+        newVnode.dom.removeAttribute(prop);
+      } else {
+        newVnode.dom.setAttribute(prop, newVnode.props[prop]);
       }
     }
-    vnodesToCleanup = [];
-  };
-
-  const addProps = (newNode: Vnode) => {
-    for (let name in newNode.props) {
-      let value = newNode.props[name];
-      if (name in newNode.props === false) {
-        return;
-      } else if (reservedWords.indexOf(name) !== -1) {
-        if (directives[name]) {
-          directives[name](value, newNode);
-        }
-      } else if (typeof value === functionstr) {
-        name = `__${name}`;
-        if (!attachedListeners[name]) {
-          (mainContainer as DomElement).addEventListener(name.slice(4), eventListener);
-          attachedListeners[name] = true;
-        }
-        (newNode.dom as DomElement)[name] = value;
-      } else if (name in (newNode.dom as DomElement) && !newNode.isSVG) {
-        // eslint-disable-next-line eqeqeq
-        if ((newNode.dom as DomElement)[name] != value) {
-          (newNode.dom as DomElement)[name] = value;
-        }
-      } else if (value !== false) {
-        (newNode.dom as DomElement).setAttribute(name, value as string);
-      }
-    }
-  };
-
-  const updateProperty = (name: string, newNode: Vnode, oldNode?: Vnode) => {
-    if (name in newNode.props) {
-      let value = newNode.props[name];
-      if (reservedWords.indexOf(name) !== -1) {
-        if (directives[name]) {
-          directives[name](value, newNode, oldNode);
-        }
-      } else if (typeof value === functionstr) {
-        name = `__${name}`;
-        if (!attachedListeners[name]) {
-          (mainContainer as DomElement).addEventListener(name.slice(4), eventListener);
-          attachedListeners[name] = true;
-        }
-        (newNode.dom as DomElement)[name] = value;
-      } else if (name in (newNode.dom as DomElement) && !newNode.isSVG) {
-        // eslint-disable-next-line eqeqeq
-        if ((newNode.dom as DomElement)[name] != value) {
-          (newNode.dom as DomElement)[name] = value;
-        }
-      } else if (!oldNode || value !== oldNode.props[name]) {
-        if (value === false) {
-          (newNode.dom as DomElement).removeAttribute(name);
-        } else {
-          (newNode.dom as DomElement).setAttribute(name, value as string);
-        }
-      }
-    }
-  };
-
+  }
   v.updateProperty = updateProperty;
 
-  let updateProps = (newNode: Vnode, oldNode?: Vnode) => {
-    for (let name in newNode.props) {
-      if (name in newNode.props === false) {
+  // Update a Vnode.dom HTMLElement with new Vnode props that are different from old Vnode props
+  function updateProperties(newVnode: Vnode & { dom: DomElement }, oldVnode?: Vnode): void {
+    for (let prop in newVnode.props) {
+      if (updateProperty(prop, newVnode, oldVnode) === false) {
         return;
       }
-      updateProperty(name, newNode, oldNode);
     }
-  };
+  }
 
-  let removeProps = (newNode: Vnode, oldNode: Vnode) => {
-    for (let name in oldNode.props) {
-      if (reservedWords.indexOf(name) === -1 && name in newNode.props === false && (oldNode === emptyNode || typeof oldNode.props[name] !== functionstr)) {
-        if (name in (newNode.dom as DomElement)) {
-          (newNode.dom as DomElement)[name] = UND;
+  function removeProperties(newVnode: Vnode & { dom: DomElement }, oldVnode: Vnode) {
+    for (let name in oldVnode.props) {
+      if (v.reservedWords.indexOf(name) === -1 && name in newVnode.props === false && typeof oldVnode.props[name] !== "function") {
+        if (name in newVnode.dom) {
+          newVnode.dom[name] = null;
         } else {
-          (newNode.dom as DomElement).removeAttribute(name);
+          newVnode.dom.removeAttribute(name);
         }
       }
     }
+  }
+
+  const callRemove = (vnode: Vnode) => {
+    for (let i = 0, l = vnode.children.length; i < l; i++) {
+      vnode.children[i] instanceof Vnode && callRemove(vnode.children[i]);
+    }
+
+    vnode.props.onremove && vnode.props.onremove(vnode);
   };
 
   let moveDom = (dom: DomElement, $parent: DomElement, oldDom: DomElement) => {
     oldDom ? $parent.replaceChild(dom, oldDom) : $parent.appendChild(dom);
   };
 
-  let updateKeyedNode = ($parent: DomElement, newNode: Vnode, newIndex: number, compareNode?: Vnode) => {
+  let updateKeyedNode = ($parent: DomElement, newVnode: Vnode & { dom: DomElement }, newIndex: number, compareNode?: Vnode & { dom: DomElement }) => {
     let oldDom = $parent.childNodes[newIndex];
 
-    // Moved or updated
-    if (compareNode) {
-      newNode.dom = compareNode.dom;
-      if (once in newNode.props || (newNode.props.onbeforeupdate && newNode.props.onbeforeupdate(newNode, compareNode) === false)) {
-        newNode.children = compareNode.children;
-        newNode.dom !== oldDom && moveDom(newNode.dom as DomElement, $parent, oldDom as DomElement);
-      } else {
-        removeProps(newNode, compareNode);
-        updateProps(newNode, compareNode);
-        newNode.dom !== oldDom && moveDom(newNode.dom as DomElement, $parent, oldDom as DomElement);
-        if (v.isMounted) {
-          newNode.props.onupdate && newNode.props.onupdate(newNode, compareNode);
-        } else {
-          newNode.props.oncreate && newNode.props.oncreate(newNode);
-        }
-        patch(newNode, compareNode);
-      }
+    if (compareNode === undefined) {
+      // Moved or updated
+      newVnode.dom = createElement(newVnode.name, newVnode.isSVG);
+      updateProperties(newVnode);
+      newVnode.dom !== oldDom && moveDom(newVnode.dom, $parent, oldDom as DomElement);
+      newVnode.props.oncreate && newVnode.props.oncreate(newVnode);
+      patch(newVnode);
     } else {
-      newNode.dom = createElement(newNode.name, newNode.isSVG);
-      addProps(newNode);
-      newNode.dom !== oldDom && moveDom(newNode.dom, $parent, oldDom as DomElement);
-      newNode.props.oncreate && newNode.props.oncreate(newNode);
-      patch(newNode);
+      newVnode.dom = compareNode.dom;
+      if ("v-once" in newVnode.props || (newVnode.props.onbeforeupdate && newVnode.props.onbeforeupdate(newVnode, compareNode) === false)) {
+        newVnode.children = compareNode.children;
+        newVnode.dom !== oldDom && moveDom(newVnode.dom as DomElement, $parent, oldDom as DomElement);
+      } else {
+        removeProperties(newVnode, compareNode);
+        updateProperties(newVnode, compareNode);
+        newVnode.dom !== oldDom && moveDom(newVnode.dom as DomElement, $parent, oldDom as DomElement);
+        newVnode.props.onupdate && newVnode.props.onupdate(newVnode, compareNode);
+        patch(newVnode, compareNode);
+      }
     }
   };
 
-  // eslint-disable-next-line complexity,sonarjs/cognitive-complexity
-  let patch = (parentNode: Vnode, oldParentNode: Vnode = emptyNode) => {
-    let newTree = isArray(parentNode.children) ? parentNode.children : [parentNode.children];
-    let oldTree = oldParentNode.children;
+  // Patch a DOM node with a new VNode tree
+  function patch(newParentVnode: Vnode & { dom: DomElement }, oldParentVnode?: Vnode & { dom: DomElement }): void {
+    let oldTree = oldParentVnode?.children || [];
+    let newTree = newParentVnode.children;
+    let oldTreeLength = oldTree.length;
 
-    current.parentVnode = parentNode;
-    current.oldParentVnode = oldParentNode;
+    current.parentVnode = newParentVnode;
+    current.oldParentVnode = oldParentVnode;
 
-    // Flatten children
+    // Flat newTree
     for (let i = 0; i < newTree.length; i++) {
-      let childVnode: unknown = newTree[i];
+      let childVnode = newTree[i];
 
       if (childVnode instanceof Vnode) {
-        (childVnode as Vnode).isSVG = parentNode.isSVG || childVnode.name === svg;
-      } else if (childVnode === NULL || childVnode === UND) {
+        childVnode.isSVG = newParentVnode.isSVG || childVnode.name === "svg";
+      } else if (childVnode === null || childVnode === undefined) {
         newTree.splice(i--, 1);
-      } else if (childVnode instanceof ComponentVnode) {
-        current.component = childVnode;
+      } else if (Array.isArray(childVnode)) {
+        newTree.splice(i--, 1, ...childVnode);
+      } else if (childVnode instanceof VnodeComponent) {
+        v.current.component = childVnode;
         newTree.splice(
           i--,
           1,
-          ("view" in (childVnode as ComponentVnode).component ? ((childVnode.component as unknown) as { view: any }).view : childVnode.component).call(
-            childVnode.component,
-            childVnode.props,
-            ...(childVnode.children as VnodeOrUnknown[])
-          )
+          ...[
+            "view" in childVnode.component
+              ? childVnode.component.view.call(childVnode.component, childVnode.props, childVnode.children)
+              : (childVnode.component as Component).call(childVnode.component, childVnode.props, childVnode.children)
+          ]
         );
-      } else if (isArray(childVnode)) {
-        newTree.splice(i--, 1, ...childVnode);
+      } else {
+        if (i > 0 && newTree[i - 1].nodeValue) {
+          newTree[i - 1].nodeValue += childVnode;
+          newTree.splice(i--, 1);
+        } else if (childVnode instanceof TextVnode === false) {
+          newTree[i] = new TextVnode(childVnode);
+        }
       }
     }
 
-    // New Tree is empty
-    if (newTree.length === 0) {
-      if (oldTree.length > 0) {
-        for (let i = oldTree.length; i--; ) {
-          oldTree[i] instanceof Vnode && callRemove(oldTree[i] as Vnode);
+    let newTreeLength = newTree.length;
+
+    // if newTree is empty, remove it
+    if (newTreeLength === 0) {
+      if (oldTreeLength > 0) {
+        for (let i = oldTreeLength; i--; ) {
+          oldTree[i] instanceof Vnode && callRemove(oldTree[i]);
         }
         // Fast node remove by setting textContent
-        (parentNode.dom as DomElement).textContent = "";
+        newParentVnode.dom.textContent = "";
       }
-
       // If the tree is keyed list and is not first render
-    } else if (oldTree.length && newTree[0] instanceof Vnode && key in newTree[0].props) {
-      let oldKeys = key in ((oldTree[0] as Vnode).props || {}) ? oldTree.map((vnode) => (vnode as Vnode).props.key) : [];
+    } else if (oldTreeLength && newTree[0] instanceof Vnode && "key" in newTree[0].props) {
+      let oldKeys = oldTree[0] instanceof Vnode && "key" in oldTree[0].props ? oldTree.map((vnode) => vnode.props.key) : [];
       let newKeys = newTree.map((vnode) => vnode.props.key);
 
       for (let i = 0, l = newKeys.length; i < l; i++) {
@@ -377,159 +376,154 @@ function valyrian(): Valyrian {
 
         // Updated: Same key
         if (key === oldKeys[i]) {
-          (oldTree[i] as Vnode).processed = true;
-          updateKeyedNode(parentNode.dom as DomElement, newNode, i, oldTree[i] as Vnode);
+          oldTree[i].processed = true;
+          updateKeyedNode(newParentVnode.dom, newNode, i, oldTree[i]);
         } else {
           let oldIndex = oldKeys.indexOf(key);
           let newIndex = i >= oldKeys.length ? -1 : i;
 
           // Moved: Key exists in old keys
           if (oldIndex !== -1) {
-            (oldTree[oldIndex] as Vnode).processed = true;
-            updateKeyedNode(parentNode.dom as DomElement, newNode, newIndex, oldTree[oldIndex] as Vnode);
+            oldTree[oldIndex].processed = true;
+            updateKeyedNode(newParentVnode.dom, newNode, newIndex, oldTree[oldIndex]);
             // Added: Key does not exists in old keys
           } else {
-            updateKeyedNode(parentNode.dom as DomElement, newNode, newIndex);
+            updateKeyedNode(newParentVnode.dom, newNode, newIndex);
           }
         }
       }
 
       // Delete unprocessed old keys
-      let l = oldTree.length;
-
-      while (l--) {
-        if (!(oldTree[l] as Vnode).processed) {
-          let oldVnode = oldTree[l];
-          oldVnode instanceof Vnode && callRemove(oldVnode);
-          (oldVnode as Vnode).dom &&
-            ((oldVnode as Vnode).dom as DomElement).parentNode &&
-            (((oldVnode as Vnode).dom as DomElement).parentNode as DomElement).removeChild((oldVnode as Vnode).dom as DomElement);
+      for (let i = oldTreeLength; i--; ) {
+        if (!oldTree[i].processed) {
+          oldTree[i] instanceof Vnode && callRemove(oldTree[i]);
+          newParentVnode.dom === oldTree[i].dom.parentNode && newParentVnode.dom.removeChild(oldTree[i].dom);
         }
       }
-
-      // Not keyed list or first render so use the simple algorithm
     } else {
-      let i = oldTree.length;
-      let l = newTree.length;
+      for (let i = 0; i < newTreeLength; i++) {
+        let childVnode = newTree[i];
+        let oldChildVnode = oldTree[i];
 
-      // Remove deleted nodes
-      while (i-- > l) {
-        let oldVnode = oldTree[i];
-        oldVnode instanceof Vnode && callRemove(oldVnode);
-        (oldVnode as Vnode).dom &&
-          ((oldVnode as Vnode).dom as DomElement).parentNode &&
-          (((oldVnode as Vnode).dom as DomElement).parentNode as DomElement).removeChild((oldVnode as Vnode).dom as DomElement);
-      }
-
-      for (i = 0; i < l; i++) {
-        let newNode = newTree[i] as Vnode | TextVnode | unknown;
-        let oldNode = oldTree[i] as Vnode | TextVnode;
-
-        // Is vnode
-        if (newNode instanceof Vnode) {
-          // Same node name
-          if (oldNode && newNode.name === (oldNode as Vnode).name) {
-            newNode.dom = (oldNode as Vnode).dom;
-            if (once in newNode.props || (newNode.props.onbeforeupdate && newNode.props.onbeforeupdate(newNode, oldNode) === false)) {
-              newNode.children = (oldNode as Vnode).children;
-            } else {
-              removeProps(newNode, oldNode as Vnode);
-              updateProps(newNode, oldNode as Vnode);
-              if (v.isMounted) {
-                newNode.props.onupdate && newNode.props.onupdate(newNode, oldNode);
-              } else {
-                newNode.props.oncreate && newNode.props.oncreate(newNode);
-              }
-              patch(newNode, oldNode as Vnode);
-            }
-
-            // Different node name
+        // if oldChildVnode is undefined, it's a new node, append it
+        if (oldChildVnode === undefined) {
+          if (childVnode instanceof Vnode) {
+            childVnode.dom = createElement(childVnode.name, childVnode.isSVG);
+            updateProperties(childVnode as Vnode & { dom: DomElement });
+            childVnode.props.oncreate && childVnode.props.oncreate(childVnode);
+            patch(childVnode as Vnode & { dom: DomElement });
           } else {
-            newNode.dom = createElement(newNode.name, newNode.isSVG);
-            addProps(newNode);
-            if (oldNode) {
-              oldNode instanceof Vnode && callRemove(oldNode);
-              (parentNode.dom as DomElement).replaceChild(newNode.dom, (parentNode.dom as DomElement).childNodes[i]);
-            } else {
-              (parentNode.dom as DomElement).appendChild(newNode.dom);
-            }
-            newNode.props.oncreate && newNode.props.oncreate(newNode);
-            patch(newNode);
+            childVnode.dom = document.createTextNode(childVnode.nodeValue);
           }
-
-          // Is TextVnode or will be
+          newParentVnode.dom.appendChild(childVnode.dom);
         } else {
-          let dom;
+          // if childVnode is Vnode, replace it with its DOM node
+          if (childVnode instanceof Vnode) {
+            if (childVnode.name === oldChildVnode.name) {
+              childVnode.dom = oldChildVnode.dom;
 
-          // If we are getting a TextVnode could be from the domToVnode method
-          let value = newNode instanceof TextVnode ? (newNode.dom as DomElement).nodeValue : String(newNode);
+              if ("v-once" in childVnode.props || (childVnode.props.onbeforeupdate && childVnode.props.onbeforeupdate(childVnode, oldChildVnode) === false)) {
+                // skip this patch
+                childVnode.children = oldChildVnode.children;
+                continue;
+              }
 
-          // Same node
-          if (oldNode instanceof TextVnode) {
-            dom = oldNode.dom;
-            // eslint-disable-next-line eqeqeq
-            if (value != (dom as DomElement).nodeValue) {
-              (dom as DomElement).nodeValue = value;
-            }
-            // Different node
-          } else {
-            dom = document.createTextNode(value as string);
-            // Old node is vnode
-            if (oldNode) {
-              callRemove(oldNode);
-              (parentNode.dom as DomElement).replaceChild(dom, oldNode.dom as DomElement);
-              // New node
+              removeProperties(childVnode as Vnode & { dom: DomElement }, oldChildVnode);
+              updateProperties(childVnode as Vnode & { dom: DomElement }, oldChildVnode);
+              if (v.isMounted) {
+                childVnode.props.onupdate && childVnode.props.onupdate(childVnode, oldChildVnode);
+              } else {
+                childVnode.props.oncreate && childVnode.props.oncreate(childVnode);
+              }
+              patch(childVnode as Vnode & { dom: DomElement }, oldChildVnode);
             } else {
-              (parentNode.dom as DomElement).appendChild(dom);
+              childVnode.dom = createElement(childVnode.name, childVnode.isSVG);
+              updateProperties(childVnode as Vnode & { dom: DomElement });
+              childVnode.props.oncreate && childVnode.props.oncreate(childVnode);
+              oldChildVnode instanceof Vnode && callRemove(oldChildVnode);
+              newParentVnode.dom.replaceChild(childVnode.dom, oldChildVnode.dom);
+              patch(childVnode as Vnode & { dom: DomElement }, oldChildVnode);
+            }
+          } else {
+            if (oldChildVnode instanceof Vnode) {
+              childVnode.dom = document.createTextNode(childVnode.nodeValue);
+              callRemove(oldChildVnode);
+              newParentVnode.dom.replaceChild(childVnode.dom, oldChildVnode.dom as DomElement);
+            } else {
+              childVnode.dom = oldChildVnode.dom;
+              // eslint-disable-next-line eqeqeq
+              if (childVnode.nodeValue != oldChildVnode.nodeValue) {
+                childVnode.dom.nodeValue = childVnode.nodeValue;
+              }
             }
           }
-          newTree[i] = new TextVnode(dom as DomElement);
         }
       }
-    }
 
-    parentNode.children = newTree;
-  };
-
-  v.update = (props: Props | null = null, ...children: VnodeOrUnknown[]) => {
-    if (mainNode) {
-      if (mountedComponent) {
-        cleanupVnodes();
-        oldMainNode = mainNode;
-        mainNode = new Vnode(mainNode.name, mainNode.props, [v(mountedComponent, props, ...children)]);
-        mainNode.dom = oldMainNode.dom;
-        mainNode.isSVG = mainNode.name === svg;
-        patch(mainNode, oldMainNode);
-        v.isMounted = true;
+      // For remaining old children: remove from DOM, garbage collect
+      for (let i = oldTreeLength - 1; i >= newTreeLength; --i) {
+        oldTree[i] instanceof Vnode && callRemove(oldTree[i]);
+        newParentVnode.dom === oldTree[i].dom.parentNode && newParentVnode.dom.removeChild(oldTree[i].dom);
       }
+    }
 
-      return v.isNode && (mainNode.dom as DomElement).innerHTML;
+    newParentVnode.children = newTree;
+  }
+
+  let mainVnode: Vnode | null = null;
+  let oldMainVnode: Vnode | null = null;
+
+  v.unMount = () => {
+    mountedComponent = emptyComponent;
+    let result = v.update();
+    v.isMounted = false;
+    mainContainer = null;
+    return result;
+  };
+
+  v.update = (props, ...children) => {
+    if (mainVnode) {
+      cleanupVnodes();
+      oldMainVnode = mainVnode;
+      mainVnode = new Vnode(mainVnode.name, mainVnode.props, [v(mountedComponent, props, ...children)]);
+      mainVnode.dom = oldMainVnode.dom;
+      mainVnode.isSVG = oldMainVnode.isSVG;
+      patch(mainVnode as Vnode & { dom: Node }, oldMainVnode as Vnode & { dom: Node });
+      v.isMounted = true;
+      if (v.isNode) {
+        return (mainVnode.dom as HTMLElement).innerHTML;
+      }
     }
   };
 
-  v.mount = (container: string, component: ValyrianComponent, props: Props | null = null, ...children: VnodeOrUnknown[]) => {
-    mainContainer = v.isNode ? createElement(container) : (document.querySelectorAll(container)[0] as DomElement);
-    mainNode = domToVnode(mainContainer) as Vnode;
-    mountedComponent = component;
+  v.mount = (container, component, props, ...children) => {
+    if (v.isMounted) {
+      v.unMount();
+    }
+
+    if (isNode) {
+      mainContainer = typeof container === "string" ? createElement(container, container === "svg") : container;
+    } else {
+      mainContainer = typeof container === "string" ? document.querySelector(container) : container;
+    }
+
+    if (mainContainer !== null) {
+      mainVnode = domToVnode(mainContainer);
+      mainVnode.isSVG = mainVnode.name === "svg";
+      oldMainVnode = mainVnode;
+      mountedComponent = component;
+    }
 
     return v.update(props, ...children);
   };
 
-  v.unMount = () => {
-    mainContainer = NULL;
-    mountedComponent = ((() => "") as unknown) as ValyrianComponent;
-    let result = v.update();
-    v.isMounted = false;
-    return result;
-  };
+  let directives: Record<string, Directive> = {};
 
-  const directives: { [x: string]: Directive } = {};
-
-  v.directive = (directive: string, handler: Directive) => {
-    let directiveName = `v-${directive}`;
-    if (reservedWords.indexOf(directiveName) === -1) {
-      reservedWords.push(directiveName);
-      directives[directiveName] = handler;
+  v.directive = (name: string, directive: Directive) => {
+    let fullName = `v-${name}`;
+    if (v.reservedWords.indexOf(fullName) === -1) {
+      v.reservedWords.push(fullName);
+      directives[fullName] = directive;
     }
   };
 
@@ -544,31 +538,31 @@ function valyrian(): Valyrian {
       vnode.name = "#text";
       vnode.children = [];
       vnode.props = {};
-      vnode.dom = (newdom as unknown) as DomElement;
+      vnode.dom = newdom as unknown as DomElement;
+      return false;
     }
   };
 
   v.directive("if", hideDirective(false));
   v.directive("unless", hideDirective(true));
-  v.directive("for", (set: unknown[], vnode: Vnode) => (vnode.children = set.map(vnode.children[0] as (value: unknown) => Function)));
-  v.directive("show", (bool: boolean, vnode: Vnode) => ((vnode.dom as { style: { display: string } }).style.display = bool ? "" : "none"));
+  v.directive("for", (set: unknown[], vnode: Vnode) => {
+    vnode.children = set.map(vnode.children[0] as (value: unknown) => Function);
+  });
+  v.directive("show", (bool: boolean, vnode: Vnode) => {
+    (vnode.dom as { style: { display: string } }).style.display = bool ? "" : "none";
+  });
   v.directive("class", (classes: { [x: string]: boolean }, vnode: Vnode) => {
     for (let name in classes) {
       (vnode.dom as DomElement).classList.toggle(name, classes[name]);
     }
   });
-  v.directive("html", (html: string, vnode: Vnode) => (vnode.children = trust(html)));
+  v.directive("html", (html: string, vnode: Vnode) => {
+    vnode.children = [trust(html)];
+  });
 
   v.newInstance = valyrian;
 
   return v;
 }
 
-(((isNode ? global : window) as unknown) as { v: Valyrian }).v = valyrian();
-
-declare global {
-  let v: Valyrian;
-  interface Window {
-    v: Valyrian;
-  }
-}
+((isNode ? global : window) as unknown as { v: Valyrian }).v = valyrian();